+++ tests/integration/redis/ConnectionTest.php
@@ -8,7 +8,7 @@
 namespace mako\tests\integration\redis;
 
 use mako\redis\Connection;
-use mako\redis\RedisException;
+use mako\redis\exceptions\RedisException;
 use mako\tests\TestCase;
 
 /**